requirements.txt
annotated-types==0.7.0
anyio==4.9.0
beautifulsoup4==4.13.3
certifi==2025.1.31
click==8.1.8
distro==1.9.0
EbookLib==0.18
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.9.0
lxml==5.3.1
Markdown==3.7
markdownify==1.1.0
openai==1.70.0
pydantic==2.11.1
pydantic_core==2.33.0
python-dotenv==1.1.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.0
